PlotLayout -> core.layout.BorderLayout
